third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbobjectstore_get2-expected.txt

This is a testharness.js-based test.
PASS IDBObjectStore.get() - key is a string  
Harness: the test ran to completion.
